ec,fs: fix clang-tidy findings from static analysis
Three categories fixed, all found via clang-tidy bugprone-* checks:

lib/fs/fs.c: remove dead store (current_de = next_de) after the ref
is already transferred to new->nm_dirent; current_de is never read
again after the break.

lib/fs/stateid.c: add comment explaining the intentional sizeof(stid)
usage.  We hash the pointer value itself (8 bytes on LP64), not the
struct it points to.  clang-tidy flags this as suspicious; the comment
silences the false positive and documents the intent.

lib/ec/mojette.c, lib/ec/mojette_codec.c: fix implicit and misplaced
widening casts in integer arithmetic used as pointer offsets.  The
pattern calloc((size_t)(P * k), ...) applies the cast after the int*int
multiplication, which can overflow on large grids.  Move the cast to the
first operand: calloc((size_t)P * k, ...) so the multiplication is done
in size_t.  Similarly grid + i * P becomes grid + (size_t)i * P.

No behavior change for current input sizes; this is a correctness
hardening against potential overflow on very large EC grids.

+++ lib/ec/mojette.c
@@ -427,7 +427,7 @@ void moj_forward(const uint64_t *__restrict__ grid, int P, int Q,
 #if defined(__aarch64__) || defined(__SSE2__) || defined(__AVX2__)
 		if (!moj_scalar_only && q == 1 && (p == 1 || p == -1)) {
 			for (int row = 0; row < Q; row++) {
-				const uint64_t *src = grid + row * P;
+				const uint64_t *src = grid + (size_t)row * P;
 				uint64_t *dst = proj->mp_bins + (off - row);
 
 #ifdef __aarch64__
